feat(storybook): theme is now being properly pulled in (#22)
