+++ .storybook/main.ts
@@ -12,7 +12,7 @@ const config: StorybookConfig = {
     '@storybook/addon-essentials',
     '@chromatic-com/storybook',
     '@storybook/addon-interactions',
-    'storybook-theme-css-vars',
+    '@storybook/addon-themes'
   ],
   framework: {
     name: '@storybook/nextjs',
