+++ package.json
@@ -59,6 +59,7 @@
     "@storybook/addon-interactions": "^8.0.5",
     "@storybook/addon-links": "^8.0.5",
     "@storybook/addon-onboarding": "^8.0.5",
+    "@storybook/addon-themes": "^8.0.6",
     "@storybook/blocks": "^8.0.5",
     "@storybook/nextjs": "^8.0.5",
     "@storybook/react": "^8.0.5",
@@ -87,7 +88,6 @@
     "lint-staged": "^15.2.2",
     "prettier": "^3.2.5",
     "storybook": "^8.0.5",
-    "storybook-theme-css-vars": "^0.0.12",
     "tsx": "^4.7.1",
     "typescript": "^5.3"
   },
